+++ node/include/cocaine/service/node/forwards.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+namespace cocaine {
+namespace api {
+
+class stream_t;
+
+}  // namespace api
+
+namespace service {
+namespace node {
+namespace slave {
+
+class id_t;
+
+}  // namespace slave
+}  // namespace node
+}  // namespace service
+}  // namespace cocaine
